move global functions to GHGRP.py as only used there
