+++ stewi/globals.py
@@ -9,11 +9,7 @@ import json
 import logging as log
 import os
 import time
-import urllib
 import copy
-import shutil
-import zipfile
-import io
 from datetime import datetime
 from pathlib import Path
 
@@ -25,7 +21,6 @@ from esupy.processed_data_mgmt import Paths, FileMeta,\
     write_df_to_file, write_metadata_to_file,\
     download_from_remote
 from esupy.dqi import get_weighted_average
-from esupy.remote import make_url_request
 from esupy.util import get_git_hash
 import stewi.exceptions
 
@@ -95,49 +90,6 @@ def config(config_path=MODULEPATH, file='config.yaml'):
     with open(path, mode='r') as f:
         configfile = yaml.load(f, Loader=yaml.FullLoader)
     return configfile
-
-
-def download_table(filepath: Path, url: str, get_time=False):
-    """Download file at url to Path if it does not exist."""
-    if not filepath.exists():
-        if url.lower().endswith('zip'):
-            r = make_url_request(url)
-            zip_file = zipfile.ZipFile(io.BytesIO(r.content))
-            zip_file.extractall(filepath)
-        elif 'xls' in url.lower() or url.lower().endswith('excel'):
-            try:
-                with urllib.request.urlopen(url) as response, open(filepath, 'wb') as out_file:
-                    shutil.copyfileobj(response, out_file)
-            except urllib.error.HTTPError:
-                log.warning(f'Error downloading {url}')
-        elif 'json' in url.lower():
-            pd.read_json(url).to_csv(filepath, index=False)
-        if get_time:
-            try:
-                retrieval_time = filepath.stat().st_ctime
-            except OSError:
-                retrieval_time = time.time()
-            return time.ctime(retrieval_time)
-    elif get_time:
-        return time.ctime(filepath.stat().st_ctime)
-
-
-def import_table(path_or_reference, get_time=False):
-    """Read and return time of csv from url or Path."""
-    try:
-        df = pd.read_csv(path_or_reference, low_memory=False)
-    except urllib.error.URLError as exception:
-        log.warning(exception.reason)
-        log.info('retrying url...')
-        time.sleep(3)
-        df = pd.read_csv(path_or_reference, low_memory=False)
-    if get_time and isinstance(path_or_reference, Path):
-        retrieval_time = path_or_reference.stat().st_ctime
-        return df, time.ctime(retrieval_time)
-    elif get_time:
-        retrieval_time = time.time()
-        return df, time.ctime(retrieval_time)
-    return df
 
 
 def aggregate(df, grouping_vars=None):
